day 3 part 2 solution
This solution uses a binary tree, well, because the problem
involves "searching" binary strings and really for learning sake.
This was tough for me to implement, tbh. I had to figure
out the correct answer manually using regex/vscode to learn
the internal structure/path the binary tree takes and compare
it to my actual implementation via println debugging. :D